feat : 생성기  - 시나리오 CRUD 추가  - Messaging / File /REST 호출 추가  - Input Adapter 추가

+++ backend/log-generator/src/main/kotlin/com/gijun/logdetect/generator/domain/model/GeneratorStatus.kt
@@ -4,5 +4,5 @@ data class GeneratorStatus(
     val running: Boolean,
     val totalSent: Long,
     val totalFailed: Long,
-    val configuredRate: Long
+    val configuredRate: Int
 )
